requirements.txt
torch
aiogram
Pillow
asyncio
torchvision==0.14.0
opencv-python==4.6.0.66